+++ src/ts/props/scrollarea.ts
@@ -19,7 +19,7 @@ export interface ScrollAreaProps extends BoxProps, StylesApiProps {
     /** Axis at which scrollbars must be rendered, `'xy'` by default */
     scrollbars?: 'x' | 'y' | 'xy' | false;
     /** Determines whether scrollbars should be offset with padding on given axis, `false` by default */
-    offsetScrollbars?: boolean | 'x' | 'y';
+    offsetScrollbars?: boolean | 'x' | 'y' | 'xy' | 'present';
     /** Defines `overscroll-behavior` of the viewport. https://developer.mozilla.org/en-US/docs/Web/CSS/overscroll-behavior */
     overscrollBehavior?: React.CSSProperties['overscrollBehavior'];
 }