+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+# [Unreleased]
+
+### Added
+
+- added `DirectionProvider` to handle RTL (right-to-left) text direction.  #650 by @AnnMarieW
+
 # 2.2.1
 
 ### Fixed