+++ CHANGELOG.md
@@ -7,7 +7,7 @@
 - added `DirectionProvider` to handle RTL (right-to-left) text direction.  #650 by @AnnMarieW
 
 ### Changed
-- Upgraded to latest Mantine (8.3.0) #650 by @AnnMarieW
+- Upgraded to latest Mantine (8.3.1) #650 by @AnnMarieW
 
 
 # 2.2.1